Fix mobile version + refactor folders
