+++ src/frontend/packages/geo-components/src/index.js
@@ -1,6 +1,4 @@
-export { default as DefaultPopupContent } from './DefaultPopupContent';
 export { default as extractContext } from './extractContext';
 export { default as LocationInput } from './LocationInput';
-export { default as MapList } from './MapList';
-export { default as MapField } from './MapField';
-export { default as QueryStringUpdater } from './QueryStringUpdater';
+export { default as MapList } from './MapList/MapList';
+export { default as MapField } from './MapField/MapField';
